Settings (#43)
* Settings are set
There is a problem to recordeo QSettingo

* Settings' base is done
It was linked directly from settings.h, exept of game controller as it stands on distance from others controllers. World window appears in center now.

* Sintax problems were fixed

* Pull request changes

* Pull request requirements're done
TODO were fixed

* Pull request ruquierements 2.0
TODO are fixed

* Pull request requirements 3.0
One crutch was fixed.

* Pull request requirements 4.0
The second crutch is fixed.

* Pull request requirements 5.0

* Pull request requirements 6.0
The name of functions is fixed.